Fix warning: cast to pointer from integer of different size (gnutls_transport_ptr_t)

+++ libmailwatch-core/mailwatch-net-conn.c
@@ -95,7 +95,7 @@ struct _XfceMailwatchNetConn
     guint port;
     const gchar *line_terminator;
 
-    gint fd;
+    glong fd;
     gint actual_port;
 
     guchar *buffer;
